mysql-innodb-cluster: no SSL by default
Don't use vault certificates by default with MySQL InnoDB Cluster.

The certificates aren't setup until you unlock the vault, which you
can't do until halfway through the charm deployment after MySQL has
already started since vault itself depends on MySQL.

When you unlock it, the certificates are created and the MySQL charm
then reconfigures and restarts MySQL which causes any OpenStack service
that is doing database migration/setup to fail and error out
permanently.

In my experience, this happens most of the time in practice.

We're not normally using this for real deployments, so add the
--mysql-innodb-cluster-ssl option to opt-in instead.