Set default constraint root-disk=20GB
Juju has an unwritten default root-disk size of 30GB which is used in
preference to the OpenStack flavors disk size when booting from volume
with root-disk-source=volume.

Set a default root-disk=20GB constraint to reduce the total storage
requirements of a default deployment, in line with what they were before
switching to booting from volumes by default in prodstack6.